Создал файл routes.py и перенес роуты туда
